refactor: Extract LLM request handler from terminal package
Move LLM interaction logic out of the terminal UI package to improve
separation of concerns and reduce coupling.

Changes:
- Move `requestResponse` function from `internal/ui/terminal/handlers.go`
  to `internal/ui/handlers/llm.go` as `RequestResponse`
- Create `handlers.AIResponse` type to replace `aiResponse`
- Update all references to use the new handler function and type
- Remove old `aiResponse` type definition from terminal types
- Clean up imports and unused files

Benefits:
- Better separation of concerns between UI and business logic
- Reusable LLM handler that can be used by other UI components
- Cleaner terminal package with focused responsibilities
- Improved testability of LLM interaction logic
